similarity/pyproject.toml
[project]
name = "similarity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bm25s>=0.2.3",
    "jax[cpu]>=0.4.35",
    "msgspec>=0.18.6",
    "numpy>=2.1.2",
    "pystemmer>=2.2.0.3",
    "sentence-transformers>=3.2.1",
    "torch>=2.5.0",
]

[tool.ruff]
line-length = 100
